+++ History.md
@@ -2,6 +2,7 @@
 ## Latest
 
 * Authenticate ArcGIS requests
+* Create `User`s for anonymous sessions for better tracking
 
 ## 0.1.2 — 2014-05-19